Chore:wrench:: #66 Modify PostingTextCoordinator
- EditingTextCoordinator -> PostingTextCoordinator로 수정
- isPostTextView로 조건 분기
